+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 the selected timeseries. The Show last year feature also works for the three date ranges.
 - The beginnings of a node.js-based graph server was added.
 - Content group tag wdfn_tng to the google analytics script.
+- Added a descriptive label and tooltip to the flood slider control.
 
 
 ### Changed